# contexts/product/core-api/docs/swagger.yaml
basePath: '{{.BasePath}}'
definitions:
  product.CreateProductCommand:
    properties:
      categoryID:
        type: integer
      initialStock:
        type: integer
      name:
        type: string
      price:
        type: string
      sku:
        type: string
    type: object
  product.DecreaseProductStockCommand:
    properties:
      stock:
        type: integer
    type: object
  product.IncreaseProductStockCommand:
    properties:
      stock:
        type: integer
    type: object
  products.Product:
    properties:
      categoryId:
        type: integer
      domainEvents:
        type: string
      finalPrice:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: string
      sku:
        type: string
      stock:
        type: integer
      vat:
        type: number
    type: object
host: '{{.Host}}'
info:
  contact: {}
  description: '{{.Description}}'
  license: {}
  title: '{{.Title}}'
  version: '{{.Version}}'
paths:
  /v1/products/:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: c
        required: true
        schema:
          $ref: '#/definitions/product.CreateProductCommand'
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/products.Product'
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
  /v1/products/decrease-stock/{sku}:
    put:
      consumes:
      - application/json
      parameters:
      - description: SKU
        in: path
        name: sku
        required: true
        type: string
      - description: body
        in: body
        name: c
        required: true
        schema:
          $ref: '#/definitions/product.DecreaseProductStockCommand'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
  /v1/products/id/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: SKU
        in: path
        name: sku
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
  /v1/products/increase-stock/{sku}:
    put:
      consumes:
      - application/json
      parameters:
      - description: SKU
        in: path
        name: sku
        required: true
        type: string
      - description: body
        in: body
        name: c
        required: true
        schema:
          $ref: '#/definitions/product.IncreaseProductStockCommand'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
swagger: "2.0"